requirements.txt
caikit-nlp-client==0.0.8
deprecation
flake8
isort
numpy
pandas
pyyaml
pylint
transformers
sentencepiece
kubernetes